+++ lang/en/theme_lexa.php
@@ -72,6 +72,9 @@ $string['mod_booking_codesdesc'] = 'Enter the <a href="https://moodle.org/plugin
 $string['navbarscrolledlayout'] = 'Scrolled navbar layout';
 $string['navbarscrolledlayoutdesc'] = 'Define the \'grid-template-columns\' CSS attribute value (<a href="https://developer.mozilla.org/Web/CSS/grid-template-columns" target="_blank">mdn web docs</a>) that should be used to define how the three columns are divided up within the space available on the navbar that is used after scrolling down the page.';
 
+$string['hideuseruserrole'] = 'Hide user role';
+$string['hideuseruserroledesc'] = 'Hide the user\'s user role on the navbar.';
+
 // Footer.
 $string['footerheading'] = 'Footer';
 $string['footerheadingsub'] = 'Footer settings';